backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.3.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
httpx<0.25